requirements.txt
click==8.1.3
decorator==5.1.1
Flask==2.2.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
jesth==0.0.5
Jinja2==3.1.2
MarkupSafe==2.1.1
pbr==5.11.0
shared==0.0.23
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
Tempita==0.5.2
Werkzeug==2.2.2
